move cache package to pkg
