+++ server/schedulers/balance_leader.go
@@ -17,7 +17,7 @@ import (
 	"strconv"
 
 	log "github.com/pingcap/log"
-	"github.com/pingcap/pd/server/cache"
+	"github.com/pingcap/pd/pkg/cache"
 	"github.com/pingcap/pd/server/core"
 	"github.com/pingcap/pd/server/schedule"
 	"go.uber.org/zap"
